+++ internal/service/mandatory_module_deletion.go
@@ -1,4 +1,4 @@
-package mandatorymodule
+package service
 
 import (
 	"context"
